fix：项目修改删除

+++ src-tauri/src/models/project.rs
@@ -1,42 +1,35 @@
-use chrono::Utc;
+use chrono::Local;
 use serde::{Deserialize, Serialize};
 use std::fs;
 use std::path::Path;
-
-#[derive(Serialize, Deserialize, Debug, Clone, Copy)]
-pub enum Layout {
-    Layout1 = 1, // 菜单和内容左右布局
-    Layout2 = 2, // 菜单和内容上下布局
-}
-
-#[derive(Serialize, Deserialize, Debug, Clone, Copy)]
-pub enum MenuMode {
-    Horizontal, // 水平菜单
-    Vertical,   // 垂直菜单
-    Inline,     // 内嵌菜单
-}
-
-#[derive(Serialize, Deserialize, Debug, Clone, Copy)]
-pub enum MenuThemeColor {
-    Dark,  // 深色主题
-    Light, // 浅色主题
+#[derive(Serialize, Deserialize, Debug)]
+pub struct Project {
+    pub id: String,                         // 项目唯一标识
+    pub name: String,                       // 项目名称
+    pub remark: String,                     // 项目备注（可选）
+    pub logo: String,                       // 项目 logo 的 URL（可选）
+    pub layout: u32,                        // 系统布局 1 2
+    pub menu_mode: String,                  // 菜单模式
+    pub menu_theme_color: String,           // 菜单主题
+    pub breadcrumb: bool,                   // 是否显示面包屑导航
+    pub tag: bool,                          // 是否显示标签页
+    pub footer: bool,                       // 是否显示页脚
+    pub system_theme_color: Option<String>, // 系统主题
+    pub created_at: String,
+    pub updated_at: String,
 }
 
 #[derive(Serialize, Deserialize, Debug)]
-pub struct Project {
-    pub id: String,                       // 项目唯一标识
-    pub name: String,                     // 项目名称
-    pub remark: String,                   // 项目备注（可选）
-    pub logo: String,             // 项目 logo 的 URL（可选）
-    pub layout: Layout,                   // 系统布局
-    pub menu_mode: MenuMode,              // 菜单模式
-    pub menu_theme_color: MenuThemeColor, // 菜单主题
-    pub breadcrumb: bool,                 // 是否显示面包屑导航
-    pub tag: bool,                        // 是否显示标签页
-    pub footer: bool,                     // 是否显示页脚
-    pub system_theme_color: Option<String>,       // 系统主题
-    pub created_at: String,
-    pub updated_at: String,
+pub struct UpdateProject {
+    pub name: String,                       // 项目名称
+    pub remark: String,                     // 项目备注（可选）
+    pub layout: u32,                        // 系统布局 1 2
+    pub menu_mode: String,                  // 菜单模式
+    pub menu_theme_color: String,           // 菜单主题
+    pub breadcrumb: bool,                   // 是否显示面包屑导航
+    pub tag: bool,                          // 是否显示标签页
+    pub footer: bool,                       // 是否显示页脚
+    pub system_theme_color: Option<String>, // 系统主题
 }
 
 #[derive(Serialize, Deserialize, Debug, Clone)]
@@ -56,17 +49,17 @@ pub struct ProjectList {
 }
 
 impl Project {
-    pub fn new(id:String , name: String, remark: String, logo: String) -> Self {
+    pub fn new(id: String, name: String, remark: String, logo: String) -> Self {
         // 生成随机并且唯一的项目 ID
-        let current_time = Utc::now().naive_utc();
+        let current_time = Local::now();
         Project {
             id,
             name,
             remark,
             logo,
-            layout: Layout::Layout1,
-            menu_mode: MenuMode::Horizontal,
-            menu_theme_color: MenuThemeColor::Dark,
+            layout: 1,
+            menu_mode: "horizontal".to_string(),
+            menu_theme_color: "light".to_string(),
             breadcrumb: true,
             tag: true,
             footer: true,
@@ -88,12 +81,31 @@ impl Project {
         serde_json::from_str(&json).unwrap()
     }
 
-    pub fn update(&self, project_path: &Path) {
-        self.save(project_path);
+    pub fn update(&self, project_path: &Path, id: String, params: UpdateProject) {
+        let mut project = Project::load(project_path);
+        project.id = id;
+        project.name = params.name;
+        project.remark = params.remark;
+        project.layout = params.layout;
+        project.menu_mode = params.menu_mode;
+        project.menu_theme_color = params.menu_theme_color;
+        project.system_theme_color = params.system_theme_color;
+        project.breadcrumb = params.breadcrumb;
+        project.tag = params.tag;
+        project.footer = params.footer;
+        project.updated_at = Local::now().format("%Y-%m-%d %H:%M:%S").to_string();
+        project.save(project_path);
     }
 
-    pub fn delete(project_path: &Path) {
-        // 删除项目目录
-        fs::remove_dir_all(project_path).unwrap();
+    pub fn delete(project_path: &Path, mode: Option<String>) {
+        // 删除项目目录 mode 如果等于 "all" 则删除项目目录
+        if let Some(mode) = mode {
+            if mode == "all" {
+                fs::remove_dir_all(project_path).unwrap();
+            }
+        } else {
+            let project_file = project_path.join("project.json");
+            fs::remove_file(project_file).unwrap();
+        }
     }
 }
